Separate emulator-dependent tests
