+++ cmd/crates/stellar-ledger/Cargo.toml
@@ -50,3 +50,6 @@ log = "0.4.21"
 once_cell = "1.19.0"
 pretty_assertions = "1.2.1"
 serial_test = "3.0.0"
+
+[features]
+emulator-tests = []
